third_party/blink/web_tests/platform/generic/css3/flexbox/css-properties-expected.txt

Tests being able to set the display to flex and inline-flex.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS flexbox.style.display is "flex"
PASS flexbox.style.display is "inline"
PASS flexbox.style.display is "inline-flex"
PASS flexbox.style.display is "inline-flex"
PASS flexbox.style.display is "block"
PASS flexitem.style.order is ""
PASS flexitem.style.order is "2"
PASS flexitem.style.order is "-1"
PASS flexitem.style.order is "0"
PASS flexitem.style.order is "0"
PASS flexitem.style.order is "0"
PASS flexitem.style.order is ""
PASS flexbox.style.justifyContent is ""
PASS window.getComputedStyle(flexbox, null).justifyContent is "normal"
PASS flexbox.style.justifyContent is ""
PASS flexbox.style.justifyContent is "flex-start"
PASS window.getComputedStyle(flexbox, null).justifyContent is "flex-start"
PASS flexbox.style.justifyContent is "flex-end"
PASS window.getComputedStyle(flexbox, null).justifyContent is "flex-end"
PASS flexbox.style.justifyContent is "center"
PASS window.getComputedStyle(flexbox, null).justifyContent is "center"
PASS flexbox.style.justifyContent is "space-between"
PASS window.getComputedStyle(flexbox, null).justifyContent is "space-between"
PASS flexbox.style.justifyContent is "space-evenly"
PASS window.getComputedStyle(flexbox, null).justifyContent is "space-evenly"
PASS flexbox.style.justifyContent is ""
PASS window.getComputedStyle(flexbox, null).justifyContent is "normal"
PASS flexbox.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignSelf is "auto"
PASS window.getComputedStyle(document.documentElement, null).alignSelf is "auto"
PASS flexbox.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignSelf is "auto"
PASS flexbox.style.alignSelf is "auto"
PASS window.getComputedStyle(flexbox, null).alignSelf is "auto"
PASS flexbox.style.alignSelf is "flex-start"
PASS window.getComputedStyle(flexbox, null).alignSelf is "flex-start"
PASS flexbox.style.alignSelf is "flex-end"
PASS window.getComputedStyle(flexbox, null).alignSelf is "flex-end"
PASS flexbox.style.alignSelf is "center"
PASS window.getComputedStyle(flexbox, null).alignSelf is "center"
PASS flexbox.style.alignSelf is "stretch"
PASS window.getComputedStyle(flexbox, null).alignSelf is "stretch"
PASS flexbox.style.alignSelf is "baseline"
PASS window.getComputedStyle(flexbox, null).alignSelf is "baseline"
PASS flexbox.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignSelf is "auto"
PASS flexbox.style.alignItems is ""
PASS flexitem.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignItems is "normal"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is ""
PASS flexitem.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignItems is "normal"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is ""
PASS flexitem.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignItems is "normal"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is "flex-start"
PASS flexitem.style.alignSelf is ""
PASS window.getComputedStyle(flexbox, null).alignItems is "flex-start"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is "flex-end"
PASS window.getComputedStyle(flexbox, null).alignItems is "flex-end"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is "center"
PASS window.getComputedStyle(flexbox, null).alignItems is "center"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is "stretch"
PASS window.getComputedStyle(flexbox, null).alignItems is "stretch"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is "baseline"
PASS window.getComputedStyle(flexbox, null).alignItems is "baseline"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is ""
PASS window.getComputedStyle(flexbox, null).alignItems is "normal"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS flexbox.style.alignItems is ""
PASS window.getComputedStyle(flexbox, null).alignItems is "normal"
PASS window.getComputedStyle(flexitem, null).alignSelf is "auto"
PASS window.getComputedStyle(detachedFlexbox, null).alignSelf is ""
PASS window.getComputedStyle(detachedFlexItem, null).alignSelf is ""
PASS flexbox.style.flexDirection is ""
PASS window.getComputedStyle(flexbox, null).flexDirection is "row"
PASS flexbox.style.flexDirection is ""
PASS flexbox.style.flexDirection is "row"
PASS window.getComputedStyle(flexbox, null).flexDirection is "row"
PASS flexbox.style.flexDirection is "row-reverse"
PASS window.getComputedStyle(flexbox, null).flexDirection is "row-reverse"
PASS flexbox.style.flexDirection is "column"
PASS window.getComputedStyle(flexbox, null).flexDirection is "column"
PASS flexbox.style.flexDirection is "column-reverse"
PASS window.getComputedStyle(flexbox, null).flexDirection is "column-reverse"
PASS flexbox.style.flexWrap is ""
PASS window.getComputedStyle(flexbox, null).flexWrap is "nowrap"
PASS flexbox.style.flexWrap is ""
PASS flexbox.style.flexWrap is "nowrap"
PASS window.getComputedStyle(flexbox, null).flexWrap is "nowrap"
PASS flexbox.style.flexWrap is "wrap"
PASS window.getComputedStyle(flexbox, null).flexWrap is "wrap"
PASS flexbox.style.flexWrap is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexWrap is "wrap-reverse"
PASS flexbox.style.flexFlow is ""
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is ""
PASS flexbox.style.flexFlow is ""
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is ""
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap"
PASS flexbox.style.flexFlow is "wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap"
PASS flexbox.style.flexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap-reverse"
PASS flexbox.style.flexFlow is "wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap-reverse"
PASS flexbox.style.flexFlow is "row"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "row"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "row nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "row nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap"
PASS flexbox.style.flexFlow is "row wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap"
PASS flexbox.style.flexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap-reverse"
PASS flexbox.style.flexFlow is "row wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row wrap-reverse"
PASS flexbox.style.flexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse nowrap"
PASS flexbox.style.flexFlow is "row-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse nowrap"
PASS flexbox.style.flexFlow is "row-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse nowrap"
PASS flexbox.style.flexFlow is "row-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse nowrap"
PASS flexbox.style.flexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse wrap"
PASS flexbox.style.flexFlow is "row-reverse wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse wrap"
PASS flexbox.style.flexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.flexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse wrap-reverse"
PASS flexbox.style.flexFlow is "column"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column nowrap"
PASS flexbox.style.flexFlow is "column"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column nowrap"
PASS flexbox.style.flexFlow is "column nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column nowrap"
PASS flexbox.style.flexFlow is "column nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column nowrap"
PASS flexbox.style.flexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column wrap"
PASS flexbox.style.flexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column wrap"
PASS flexbox.style.flexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column wrap-reverse"
PASS flexbox.style.flexFlow is "column wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column wrap-reverse"
PASS flexbox.style.flexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse nowrap"
PASS flexbox.style.flexFlow is "column-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse nowrap"
PASS flexbox.style.flexFlow is "column-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse nowrap"
PASS flexbox.style.flexFlow is "column-reverse nowrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse nowrap"
PASS flexbox.style.flexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse wrap"
PASS flexbox.style.flexFlow is "column-reverse wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse wrap"
PASS flexbox.style.flexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.flexFlow is "column-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column-reverse wrap-reverse"
PASS flexbox.style.flexFlow is ""
PASS flexbox.style.flexFlow is ""
PASS flexbox.style.flexFlow is ""
PASS flexbox.style.flexFlow is ""
PASS window.getComputedStyle(flexbox, null).flexFlow is "row nowrap"
PASS flexbox.style.flexFlow is "column"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column nowrap"
PASS flexbox.style.flexFlow is "column wrap"
PASS window.getComputedStyle(flexbox, null).flexFlow is "column wrap"
PASS flexbox.style.flexFlow is "row-reverse wrap-reverse"
PASS window.getComputedStyle(flexbox, null).flexFlow is "row-reverse wrap-reverse"
PASS foundFlexFlow is false
PASS flexbox.style.alignContent is ""
PASS window.getComputedStyle(flexbox, null).alignContent is "normal"
PASS flexbox.style.alignContent is "flex-start"
PASS window.getComputedStyle(flexbox, null).alignContent is "flex-start"
PASS flexbox.style.alignContent is "flex-end"
PASS window.getComputedStyle(flexbox, null).alignContent is "flex-end"
PASS flexbox.style.alignContent is "center"
PASS window.getComputedStyle(flexbox, null).alignContent is "center"
PASS flexbox.style.alignContent is "space-between"
PASS window.getComputedStyle(flexbox, null).alignContent is "space-between"
PASS flexbox.style.alignContent is "space-around"
PASS window.getComputedStyle(flexbox, null).alignContent is "space-around"
PASS flexbox.style.alignContent is "space-evenly"
PASS window.getComputedStyle(flexbox, null).alignContent is "space-evenly"
PASS flexbox.style.alignContent is "stretch"
PASS window.getComputedStyle(flexbox, null).alignContent is "stretch"
PASS flexbox.style.alignContent is ""
PASS window.getComputedStyle(flexbox, null).alignContent is "normal"
PASS flexbox.style.alignContent is ""
PASS window.getComputedStyle(flexbox, null).alignContent is "normal"
PASS successfullyParsed is true

TEST COMPLETE
